+++ admin/views/components/file-browser/utils/append-path.js
